requirements.txt
# cffi==1.4.2
cffi
# cryptography==1.1.2
# cryptography
Django==1.9.2
django-jquery==1.9.1
django-lockdown==1.2
enum34==1.1.2
gspread==0.3.0
httplib2==0.9.2
idna==2.0
ipaddress==1.0.16
MySQL-python==1.2.5
# MySQL-python
# mysql-connector-python
# mysqlclient
oauth2client==1.5.2
ptvsd
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
#pyOpenSSL==0.15.1
pyOpenSSL
# python-mysqldb
requests==2.7.0
rsa==3.2.3
six==1.10.0
sqlparse==0.1.18
xlrd==0.9.3
django-debug-toolbar==1.6
# django-debug-toolbar
django-extensions==1.7.6
django-chartit==0.2.9
